fix(db): update domains types

+++ db/01_schema/02_domains/electric_combined_bank_power_units.sql
@@ -1,6 +1,6 @@
 -- TABLE: electric_combined_bank_power_units
 CREATE TABLE IF NOT EXISTS domains.electric_combined_bank_power_units (
-    code smallint PRIMARY KEY,
+    code bigint PRIMARY KEY,
     description varchar(64),
     details varchar(2000)
 );
