+++ db/01_schema/02_domains/electric_wire_diameter_in.sql
@@ -1,6 +1,6 @@
 -- TABLE: electric_wire_diameter_in
 CREATE TABLE IF NOT EXISTS domains.electric_wire_diameter_in (
-    code smallint PRIMARY KEY,
+    code numeric PRIMARY KEY,
     description varchar(64),
     details varchar(2000)
 );
